Game loop functionality
added the logic to enable the player to win and also start a new game once they have won.
Used react confetti to add some visual effects to the win screen.
added some conditions to change the button text and the onClick function used by the button dependant on if the player has won or not.